requirements.txt
requests
pandas
ortools
flask
gunicorn
python-dotenv 